java-cxf-client: rework from java-cxf-cdi-client
CDI turns out to be unnecessary and cumbersome for the API client. Better to incorporate the simple implementation into your own CDI setup for configuration if you need it.

Refocuses the client on the high-level interface that includes our exceptions, rather than the previous approach of focussing on the JAX-RS interfaces. The JAX-RS interfaces are now in a separate package.